requirements.txt
streamlit
pycryptodome
cryptography
pillow
imagehash
pycipher
